Self-review, resolved some of todos
